# app/apidocs/apiDoc.v1.0.1.json
{
    "openapi": "3.0.1",
    "info": {
        "version" : "1.0.1",
        "title": "Sale Report Manager APIs Document",
        "description": "your description here",
        "termsOfService": "",
        "contact": {
            "name": "Sales Report Manager",
            "email": "parveen@acewebx.com",
            "url": "https://acewebx.com/"
        }
    },
    "servers": [
        {
            "url": "https://sales-report-manager.herokuapp.com/api/v1.0.1",
            "description": "Production server"            
        },
        {
            "url": "http://localhost:8444/api/v1.0.1",
            "description": "Development server"            
        }
    ],
...
    "paths": {
...
      "/auth/loginWithAccessToken": {
        "post": {
          "tags": ["Auth"],
          "summery": "Login with Access Token",
          "description": "Get user object",
          "operationId": "loginWithAccessToken",          
          "responses": {
            "400": {
              "description": "Missing parameters"              
            }
          },
          "security": [
            {
              "jwt": []
            }
          ]
        }
      },
...
    },
    "components": {
        "securitySchemes": {
          "jwt": {
            "name": "x-auth-token",
            "type": "apiKey",
            "scheme": "bearer",
            "in": "header",
            "bearerFormat": "JWT"
          }
        }
    }
}